+++ R/generate_all_datasets_to_files.R
@@ -22,6 +22,7 @@ generate_all_dummy_data_to_files<-function(
     person_level_data_version="R12v1",
     kidney_level_data_version = "R12v1",
     birth_mother_level_data_version="R12v1",
+    kanta_medication_delivery_level_data_version="R12v1",
     vision_level_data_version="R12v1",
     n_patients_minimum = 100,
     per_patients_service_sector = 0.99,
@@ -79,7 +80,7 @@ generate_all_dummy_data_to_files<-function(
   #
   ParallelLogger::logInfo("Generate kidney level data")
 
-  kidney_extended <- generate_dummy_kidney_register_data(
+  kidney_data <- generate_dummy_kidney_register_data(
     kidney_level_data_version = kidney_level_data_version,
     n_patients_minimum = n_patients_minimum,
     seed = seed
@@ -90,8 +91,20 @@ generate_all_dummy_data_to_files<-function(
   #
   ParallelLogger::logInfo("Generate birth mother level data")
 
-  birth_mother_extended <- generate_dummy_mothers_birth_register_data(
+  birth_mother_data <- generate_dummy_mothers_birth_register_data(
     birth_mother_level_data_version = birth_mother_level_data_version,
+    n_patients_minimum = n_patients_minimum,
+    seed = seed,
+    service_sector_data = service_sector_data
+  )
+
+  #
+  # generate kanta medication delivery level data
+  #
+  ParallelLogger::logInfo("Generate kanta medication delivery level data")
+
+  kanta_medication_delivery_data <- generate_dummy_kanta_medication_delivery_data(
+    kanta_medication_delivery_level_data_version = kanta_medication_delivery_level_data_version,
     n_patients_minimum = n_patients_minimum,
     seed = seed,
     service_sector_data = service_sector_data
@@ -103,7 +116,7 @@ generate_all_dummy_data_to_files<-function(
   #
   ParallelLogger::logInfo("Generate vision level data")
 
-  vision_extended <- generate_dummy_vision_register_data(
+  vision_data <- generate_dummy_vision_register_data(
     vision_level_data_version = vision_level_data_version,
     n_patients_minimum = n_patients_minimum,
     seed = seed
@@ -119,14 +132,17 @@ generate_all_dummy_data_to_files<-function(
   file_name <- stringr::str_c("dummy_minimum_extended_", person_level_data_version, ".txt" )
   minimum_extended |> readr::write_tsv(file.path(output_folder, file_name))
 
-  file_name <- stringr::str_c("dummy_kidney_extended_", vision_level_data_version, ".txt" )
-  kidney_extended |> readr::write_tsv(file.path(output_folder, file_name))
+  file_name <- stringr::str_c("dummy_kidney_", kidney_level_data_version, ".txt" )
+  kidney_data |> readr::write_tsv(file.path(output_folder, file_name))
 
-  file_name <- stringr::str_c("dummy_birth_mother_extended_", birth_mother_level_data_version, ".txt" )
-  birth_mother_extended |> readr::write_tsv(file.path(output_folder, file_name))
+  file_name <- stringr::str_c("dummy_birth_mother_", birth_mother_level_data_version, ".txt" )
+  birth_mother_data |> readr::write_tsv(file.path(output_folder, file_name))
 
-  file_name <- stringr::str_c("dummy_vision_extended_", vision_level_data_version, ".txt" )
-  vision_extended |> readr::write_tsv(file.path(output_folder, file_name))
+  file_name <- stringr::str_c("dummy_kanta_medication_delivery_", kanta_medication_delivery_level_data_version, ".txt" )
+  kanta_medication_delivery_data |> readr::write_tsv(file.path(output_folder, file_name))
+
+  file_name <- stringr::str_c("dummy_vision_", vision_level_data_version, ".txt" )
+  vision_data |> readr::write_tsv(file.path(output_folder, file_name))
 
   ParallelLogger::logInfo("Saved data")
   ParallelLogger::unregisterLogger(logger)